Introduced dedicated NakadiRecordResource class
