+++ core-common/src/main/java/org/zalando/nakadi/domain/NakadiRecord.java
@@ -1,47 +1,8 @@
 package org.zalando.nakadi.domain;
 
-import org.zalando.nakadi.plugin.api.authz.AuthorizationAttribute;
-import org.zalando.nakadi.plugin.api.authz.AuthorizationService;
-import org.zalando.nakadi.plugin.api.authz.Resource;
-
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-
-public class NakadiRecord implements Resource<NakadiRecord> {
+public class NakadiRecord {
 
     public static final String HEADER_FORMAT = new String(new byte[]{0});
-
-    // todo maybe use a dedicated resource object
-    @Override
-    public String getName() {
-        return metadata.getEid();
-    }
-
-    @Override
-    public String getType() {
-        return ResourceImpl.EVENT_RESOURCE;
-    }
-
-    @Override
-    public Optional<List<AuthorizationAttribute>> getAttributesForOperation(
-            final AuthorizationService.Operation operation) {
-        if (operation == AuthorizationService.Operation.WRITE) {
-            return Optional.ofNullable(owner).map(AuthorizationAttributeProxy::new).map(Collections::singletonList);
-        }
-        return Optional.empty();
-    }
-
-    @Override
-    public NakadiRecord get() {
-        return this;
-    }
-
-    @Override
-    public Map<String, List<AuthorizationAttribute>> getAuthorization() {
-        return null;
-    }
 
     public enum Format {
         AVRO(new byte[]{0});
