+++ core-services/src/main/java/org/zalando/nakadi/service/AuthorizationValidator.java
@@ -7,6 +7,7 @@ import org.zalando.nakadi.cache.EventTypeCache;
 import org.zalando.nakadi.domain.BatchItem;
 import org.zalando.nakadi.domain.EventType;
 import org.zalando.nakadi.domain.NakadiRecord;
+import org.zalando.nakadi.domain.NakadiRecordResource;
 import org.zalando.nakadi.domain.Subscription;
 import org.zalando.nakadi.exceptions.runtime.AccessDeniedException;
 import org.zalando.nakadi.exceptions.runtime.ForbiddenOperationException;
@@ -98,7 +99,8 @@ public class AuthorizationValidator {
         } catch (AuthorizationInvalidException e) {
             throw new UnprocessableEntityException(e.getMessage());
         } catch (PluginException e) {
-            throw new ServiceTemporarilyUnavailableException("Error calling authorization plugin: "+ e.getMessage(), e);
+            throw new ServiceTemporarilyUnavailableException("Error calling authorization plugin: "
+                    + e.getMessage(), e);
         }
     }
 
@@ -127,10 +129,11 @@ public class AuthorizationValidator {
             return;
         }
         try {
+            final var nakadiRecordResource = new NakadiRecordResource(nakadiRecord);
             final boolean authorized = authorizationService.isAuthorized(
-                    AuthorizationService.Operation.WRITE, nakadiRecord);
+                    AuthorizationService.Operation.WRITE, nakadiRecordResource);
             if (!authorized) {
-                throw new AccessDeniedException(AuthorizationService.Operation.WRITE, nakadiRecord);
+                throw new AccessDeniedException(AuthorizationService.Operation.WRITE, nakadiRecordResource);
             }
         } catch (final PluginException ex) {
             throw new ServiceTemporarilyUnavailableException("Error while checking authorization", ex);
@@ -161,7 +164,8 @@ public class AuthorizationValidator {
                 }
             }
         } catch (final PluginException e) {
-            throw new ServiceTemporarilyUnavailableException("Error calling authorization plugin: "+ e.getMessage(), e);
+            throw new ServiceTemporarilyUnavailableException("Error calling authorization plugin: "
+                    + e.getMessage(), e);
         }
     }
 
@@ -172,7 +176,8 @@ public class AuthorizationValidator {
                 throw new AccessDeniedException(AuthorizationService.Operation.VIEW, resource);
             }
         } catch (final PluginException e) {
-            throw new ServiceTemporarilyUnavailableException("Error calling authorization plugin: "+ e.getMessage(), e);
+            throw new ServiceTemporarilyUnavailableException("Error calling authorization plugin: "
+                    + e.getMessage(), e);
         }
     }
 
@@ -236,7 +241,8 @@ public class AuthorizationValidator {
                 throw new AccessDeniedException(AuthorizationService.Operation.READ, resource);
             }
         } catch (final PluginException e) {
-            throw new ServiceTemporarilyUnavailableException("Error calling authorization plugin: "+ e.getMessage(), e);
+            throw new ServiceTemporarilyUnavailableException("Error calling authorization plugin: "
+                    + e.getMessage(), e);
         }
     }
